+++ androidsample/src/androidTest/java/io/appflate/restmock/androidsample/TestApplication.java
@@ -16,6 +16,7 @@
 
 package io.appflate.restmock.androidsample;
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import io.appflate.restmock.RESTMockServer;
 import io.appflate.restmock.androidsample.di.AppModule;
 import io.appflate.restmock.androidsample.di.DaggerAppComponent;
@@ -24,7 +25,9 @@ import io.appflate.restmock.androidsample.di.DaggerAppComponent;
  * Created by andrzejchm on 23/04/16.
  */
 public class TestApplication extends SampleApplication {
+
     @Override
+    @SuppressFBWarnings("ST_WRITE_TO_STATIC_FROM_INSTANCE_METHOD")
     protected void setupGraph() {
         //here I'm supplying the AppModule with RESTMock's url instead of github's API url
         appComponent = DaggerAppComponent.builder()