+++ core/src/main/java/io/appflate/restmock/RESTMockServer.java
@@ -16,8 +16,8 @@
 
 package io.appflate.restmock;
 
-import com.squareup.okhttp.mockwebserver.MockWebServer;
-import com.squareup.okhttp.mockwebserver.RecordedRequest;
+import okhttp3.mockwebserver.MockWebServer;
+import okhttp3.mockwebserver.RecordedRequest;
 
 import org.hamcrest.Matcher;
 
@@ -25,7 +25,6 @@ import java.io.IOException;
 
 import io.appflate.restmock.logging.NOOpLogger;
 import io.appflate.restmock.logging.RESTMockLogger;
-import io.appflate.restmock.utils.RequestMatcher;
 
 import static io.appflate.restmock.utils.RequestMatchers.isDELETE;
 import static io.appflate.restmock.utils.RequestMatchers.isGET;
@@ -37,14 +36,19 @@ import static org.hamcrest.core.AllOf.allOf;
 
 public class RESTMockServer {
 
-    private static MockWebServer mockWebServer;
-    private static MatchableCallsRequestDispatcher dispatcher;
+    public static final String RESPONSE_NOT_MOCKED = "NOT MOCKED";
+    public static final String MORE_THAN_ONE_RESPONSE_ERROR = "There are more than one response matching this request: ";
+    static MockWebServer mockWebServer;
+    static MatchableCallsRequestDispatcher dispatcher;
     private static String serverBaseUrl;
     private static RESTMockFileParser RESTMockFileParser;
     static RESTMockLogger logger;
 
-    public static void init(RESTMockFileParser RESTMockFileParser,
+    public synchronized static void init(RESTMockFileParser RESTMockFileParser,
                             RESTMockLogger logger) throws IOException {
+        if(RESTMockServer.mockWebServer != null) {
+            RESTMockServer.shutdown();
+        }
         RESTMockServer.mockWebServer = new MockWebServer();
         if (logger == null) {
             RESTMockServer.logger = new NOOpLogger();
@@ -56,6 +60,7 @@ public class RESTMockServer {
         RESTMockServer.mockWebServer.setDispatcher(dispatcher);
         RESTMockServer.mockWebServer.start();
         RESTMockServer.serverBaseUrl = mockWebServer.url("/").toString();
+        RequestsVerifier.init(dispatcher);
 
         RESTMockServer.RESTMockFileParser = RESTMockFileParser;
         RESTMockServer.logger.log("## RESTMock successfully started!\turl: " + RESTMockServer.serverBaseUrl);
@@ -81,10 +86,18 @@ public class RESTMockServer {
     }
 
     /**
-     * removes all matchable calls stored in this {@code RESTMockServer}
+     * Use {@link #reset() instead}
      */
+    @Deprecated
     public static void removeAllMatchableCalls() {
+        reset();
+    }
+    /**
+     * removes all mocks stored in this {@code RESTMockServer} as well as all history requests
+     */
+    public static void reset() {
         dispatcher.removeAllMatchableCalls();
+        dispatcher.clearHistoricalRequests();
     }
 
     /**
@@ -132,7 +145,7 @@ public class RESTMockServer {
      * @param requestMatcher matcher to match a GET request
      * @return {@code MatchableCall} that will match GET requests along with {@code requestMatcher}
      */
-    public static MatchableCall whenGET(RequestMatcher requestMatcher) {
+    public static MatchableCall whenGET(Matcher<RecordedRequest> requestMatcher) {
         return RESTMockServer.whenRequested(allOf(isGET(), requestMatcher));
     }
 
@@ -142,7 +155,7 @@ public class RESTMockServer {
      * @param requestMatcher matcher to match a POST request
      * @return {@code MatchableCall} that will match POST requests along with {@code requestMatcher}
      */
-    public static MatchableCall whenPOST(RequestMatcher requestMatcher) {
+    public static MatchableCall whenPOST(Matcher<RecordedRequest>  requestMatcher) {
         return RESTMockServer.whenRequested(allOf(isPOST(), requestMatcher));
     }
 
@@ -152,7 +165,7 @@ public class RESTMockServer {
      * @param requestMatcher matcher to match a PUT request
      * @return {@code MatchableCall} that will match PUT requests along with {@code requestMatcher}
      */
-    public static MatchableCall whenPUT(RequestMatcher requestMatcher) {
+    public static MatchableCall whenPUT(Matcher<RecordedRequest>  requestMatcher) {
         return RESTMockServer.whenRequested(allOf(isPUT(), requestMatcher));
     }
 
@@ -162,7 +175,7 @@ public class RESTMockServer {
      * @param requestMatcher matcher to match a PATCH request
      * @return {@code MatchableCall} that will match PATCH requests along with {@code requestMatcher}
      */
-    public static MatchableCall whenPATCH(RequestMatcher requestMatcher) {
+    public static MatchableCall whenPATCH(Matcher<RecordedRequest>  requestMatcher) {
         return RESTMockServer.whenRequested(allOf(isPATCH(), requestMatcher));
     }
 
@@ -172,7 +185,7 @@ public class RESTMockServer {
      * @param requestMatcher matcher to match a DELETE request
      * @return {@code MatchableCall} that will match DELETE requests along with {@code requestMatcher}
      */
-    public static MatchableCall whenDELETE(RequestMatcher requestMatcher) {
+    public static MatchableCall whenDELETE(Matcher<RecordedRequest>  requestMatcher) {
         return RESTMockServer.whenRequested(allOf(isDELETE(), requestMatcher));
     }
 
@@ -186,5 +199,14 @@ public class RESTMockServer {
      */
     public static MatchableCall whenRequested(Matcher<RecordedRequest> requestMatcher) {
         return new MatchableCall(RESTMockServer.RESTMockFileParser, requestMatcher, dispatcher);
+    }
+
+    /**
+     * Shuts down the instance of RESTMockServer
+     * @throws IOException if something goes wrong while stopping
+     */
+    public static void shutdown() throws IOException {
+        reset();
+        mockWebServer.shutdown();
     }
 }