+++ core/src/main/java/io/appflate/restmock/utils/RequestMatcher.java
@@ -16,7 +16,7 @@
 
 package io.appflate.restmock.utils;
 
-import com.squareup.okhttp.mockwebserver.RecordedRequest;
+import okhttp3.mockwebserver.RecordedRequest;
 
 import org.hamcrest.Description;
 import org.hamcrest.TypeSafeMatcher;
@@ -25,7 +25,6 @@ import org.hamcrest.TypeSafeMatcher;
 /**
  * <p>A RequestMatcher is an extension of {@link org.hamcrest.TypeSafeMatcher} making it easier to specify
  * the matcher without the need of implementing {@link #describeTo(Description)} method.</p>
- * <p>
  * <p>See {@link org.hamcrest.Matcher} for more info about hamcrest's matchers</p>
  */
 public abstract class RequestMatcher extends TypeSafeMatcher<RecordedRequest> {